Integrate SPL as a bundled component.
Remove autotool tests for SPL components add them directly to the
autoconf and automake templates for ZFS.

The headers bundled with SPL are incompatible with those bundled with
ZFS and cannot be merged without causing conflicts in sys/types.h and
sys/sysmacros.h inclusions. The next development step is to merge the
include.spl/ headers from SPL into the existing include/ tree.